remove 'impute' word,add feature with list of projectId to Trello settings,also for GoogleDrive;Correct grammar mistakes,Correct display  status methods for all services,Correct README,

+++ src/main/java/tk/microlms/accessmanager/cli/Main.java
@@ -1,8 +1,6 @@
 package tk.microlms.accessmanager.cli;
 
-import tk.microlms.accessmanager.model.GitlabRolePermission;
-import tk.microlms.accessmanager.model.GitlabUser;
-import tk.microlms.accessmanager.model.TrelloUser;
+import tk.microlms.accessmanager.model.*;
 import tk.microlms.accessmanager.service.*;
 
 import java.util.List;
@@ -10,9 +8,11 @@ import java.util.Scanner;
 
 public class Main {
     public static void main(String[] args) throws Exception {
-        final int EXIT_STATUS = 0;
-        Scanner scanner = new Scanner(System.in);
+        ConfigurationService configurationService = new ConfigurationService();
         FilePathService filePathService = new FilePathService();
+        ConfigurationCli configurationCli = new ConfigurationCli();
+        Scanner scanner = new Scanner(System.in);
+        final int EXIT_STATUS = 0;
 
         if (!filePathService.checkConfFile()) {
             String errorMassage = String.format("File 'configuration.json' is found neither in the current directory (%s)" +
@@ -24,8 +24,18 @@ public class Main {
                 result = result.toLowerCase();
 
                 if (result.equals("y")) {
-                    ConfigurationCli configurationCli = new ConfigurationCli();
-                    configurationCli.newConfigurations();
+
+                    GoogleDriveClient googleDriveClientConf = configurationCli.newGoogleDriveClientConf();
+                    configurationService.saveGoogleDriveClientToFile(googleDriveClientConf);
+                    GoogledriveUser googledriveUserConf = configurationCli.newGoogleDriveConf();
+                    List<GitlabUser> gitlabUsersConf = configurationCli.newGitLabConf();
+                    TrelloUser trelloUserConf = configurationCli.newTrelloConf();
+
+                    configurationService
+                        .newConf(trelloUserConf,
+                            gitlabUsersConf,
+                            googledriveUserConf,
+                            googleDriveClientConf);
                     break;
                 }
                 if (result.equals("n")) {
@@ -45,9 +55,9 @@ public class Main {
         GoogledriveService googleDriveService = new GoogledriveService(googleFileId);
 
         System.out.println(
-            "Add to GitlabUser repositories,GoogledriveUser directory,TrelloUser board impute 'a'" +
-                "\nGet status impute 's' " +
-                "\nQuit the program impute 'q'" + "\nDelete a member impute'd'");
+            "\nAdd to GitlabUser repositories,GoogledriveUser directory,TrelloUser board enter 'a'" +
+                "\nGet status enter 's' " +
+                "\nQuit the program enter 'q'" + "\nDelete a member enter 'd'");
 
         while (true) {
             String result = scanner.next();
@@ -58,56 +68,71 @@ public class Main {
             }
 
             if (result.equals("a")) {
-                System.out.println("impute email:");
-                String email = scanner.next();
-                email = email.toLowerCase();
+                while (true) {
+                    System.out.println("Enter email:");
+                    String email = scanner.next();
+                    if (configurationCli.isValid(email)) {
+                        email = email.toLowerCase();
 
-                trelloService.addToTrello(email);
-                googleDriveService.addToGoogleDrive(email);
+                        trelloService.addToTrello(email);
+                        System.out.println("add to Trello");
 
-                List<GitlabUser> gitlabUsers = ReaderService.readGitLabConf();
-                for (int i = 0; i < gitlabUsers.size(); i++) {
+                        googleDriveService.addToGoogleDrive(email);
+                        System.out.println("add to GoogleDrive");
 
-                    String tempUserName = ReaderService.readGitLabConf().get(i).getUsername();
-                    String tempPass = ReaderService.readGitLabConf().get(i).getPass();
-                    String tempProjectId = ReaderService.readGitLabConf().get(i).getProjectId();
+                        List<GitlabUser> gitlabUsers = ReaderService.readGitLabConf();
+                        for (int i = 0; i < gitlabUsers.size(); i++) {
 
-                    GitlabService gitLabService = new GitlabService(tempUserName, tempPass);
-                    gitLabService.addToGitLab(Integer.valueOf(tempProjectId), email,
-                        GitlabRolePermission.DEVELOPER.getLevel());
+                            String tempUserName = ReaderService.readGitLabConf().get(i).getUsername();
+                            String tempPass = ReaderService.readGitLabConf().get(i).getPass();
+                            String tempProjectId = ReaderService.readGitLabConf().get(i).getProjectId();
+
+                            GitlabService gitLabService = new GitlabService(tempUserName, tempPass);
+                            gitLabService.addToGitLab(Integer.valueOf(tempProjectId), email,
+                                GitlabRolePermission.DEVELOPER.getLevel());
+                        }
+                        System.out.println("add to GitLab");
+                        break;
+                    } else {
+                        System.out.println("Email is not valid...\n");
+                    }
                 }
-                System.out.println("add to GitLab");
             }
             if (result.equals("d")) {
-                System.out.println("impute email:");
-                String email = scanner.next();
-                email = email.toLowerCase();
+                while (true) {
+                    System.out.println("Enter email:");
+                    String email = scanner.next();
 
-                trelloService.deleteMember(email);
-                System.out.println("delete from trello");
+                    if (configurationCli.isValid(email)) {
+                        email = email.toLowerCase();
 
-                googleDriveService.delete(email);
-                System.out.println("delete from GoogleDrive");
+                        trelloService.deleteMember(email);
+                        System.out.println("delete from trello");
 
-                List<GitlabUser> gitlabUsers = ReaderService.readGitLabConf();
+                        googleDriveService.delete(email);
+                        System.out.println("delete from GoogleDrive");
 
-                for (int i = 0; i < gitlabUsers.size(); i++) {
-                    String tempUserName = ReaderService.readGitLabConf().get(i).getUsername();
-                    String tempPass = ReaderService.readGitLabConf().get(i).getPass();
-                    String tempProjectId = ReaderService.readGitLabConf().get(i).getProjectId();
+                        List<GitlabUser> gitlabUsers = ReaderService.readGitLabConf();
 
-                    GitlabService gitLabService = new GitlabService(tempUserName, tempPass);
-                    gitLabService.delete(Integer.valueOf(tempProjectId), email);
+                        for (int i = 0; i < gitlabUsers.size(); i++) {
+                            String tempUserName = ReaderService.readGitLabConf().get(i).getUsername();
+                            String tempPass = ReaderService.readGitLabConf().get(i).getPass();
+                            String tempProjectId = ReaderService.readGitLabConf().get(i).getProjectId();
+
+                            GitlabService gitLabService = new GitlabService(tempUserName, tempPass);
+                            gitLabService.delete(Integer.valueOf(tempProjectId), email);
+                        }
+                        System.out.println("delete from GitLab");
+                        break;
+                    } else {
+                        System.out.println("Email is not valid...\n");
+                    }
                 }
-                System.out.println("delete from GitLab");
             }
 
             if (result.equals("s")) {
                 System.out.println("TrelloUser status:");
                 trelloService.getStatus();
-
-                System.out.println("\nGoogledriveUser status:");
-                googleDriveService.getStatus();
 
                 System.out.println("\nGitlabUser status:");
 
@@ -122,6 +147,9 @@ public class Main {
                     System.out.println("\nProjectId:" + tempProjectId);
                     gitLabService.getStatus(tempProjectId);
                 }
+
+                System.out.println("\nGoogledriveUser status:");
+                googleDriveService.getStatus();
             }
         }
     }
